- Fixed bug #13627: Files larger than 16MB does not get stored when using clustering.   (Merged from trunk (4.1.0alpha1) rev. 22359)
git-svn-id: file:///home/patrick.allaert/svn-git/ezp-repo/ezpublish/stable/4.0@22360 a01eee8c-daba-0310-acae-fa49f3023285

+++ kernel/classes/clusterfilehandlers/dbbackends/mysql.php
@@ -52,7 +52,7 @@ CREATE TABLE ezdbfile (
 
 CREATE TABLE ezdbfile_data (
   name_hash VARCHAR(34)   NOT NULL DEFAULT '',
-  offset    MEDIUMINT(11) UNSIGNED NOT NULL,
+  offset    INT(11) UNSIGNED NOT NULL,
   filedata  BLOB          NOT NULL,
   PRIMARY KEY (name_hash, offset),
   CONSTRAINT ezdbfile_fk1 FOREIGN KEY (name_hash) REFERENCES ezdbfile (name_hash) ON DELETE CASCADE
